+++ probe-rs/src/architecture/arm/core/m0.rs
@@ -531,6 +531,14 @@ impl<'probe> CoreInterface for M0<'probe> {
     fn status(&mut self) -> Result<crate::core::CoreStatus, Error> {
         let dhcsr = Dhcsr(self.memory.read_word_32(Dhcsr::ADDRESS)?);
 
+        if dhcsr.s_lockup() {
+            log::warn!("The core is in locked up status as a result of an unrecoverable exception");
+
+            self.state.current_state = CoreStatus::LockedUp;
+
+            return Ok(CoreStatus::LockedUp);
+        }
+
         if dhcsr.s_sleep() {
             // Check if we assumed the core to be halted
             if self.state.current_state.is_halted() {